[Parse] The Big Split
Separated the architecture module into multiple pieces to be a bit more
organized. I decided to split the files amongst the top-level tags in
the architecture description file.
